Fix parsing of OpenSSL self-signed PSS certs
When generating an issuer in OpenSSL using:

> openssl  req -x509 -newkey rsa-pss -keyout rsassaPss.key \
>  -out rsassaPss.crt -subj /CN=localhost -nodes -batch \
>  -pkeyopt rsa_keygen_bits:2048 -sha256

the subsequent certificate will not be accepted by Go's
ParseCertificate(...) because the PSS can have an optional NULL
parameter which OpenSSL elides (versus RSAEncryption which appears
to mandate this parameter as non-optional but NULL).

During subsequent verification (CheckSignatureFrom(...) in the test
suite), this certificate again fails due to strict PSS verification
requirements: the saltLength MUST equal the hash length. However,
OpenSSL defaults here to the maximum salt length (222), breaking Go's
verification (which uses rsa.PSSSaltLengthEqualsHash). Switch this to
verifying the salt length is at least the hash size and using the
alternative rsa.PSSSaltLengthAuto constant on verification instead.

Signed-off-by: Alexander Scheel <alex.scheel@hashicorp.com>